Move Calender component to own file
