+++ src/app/governance/components/CalendarCards.tsx
@@ -16,9 +16,7 @@ import { Heading } from '@/components/Heading'
 import { TagGroupContainer } from '@/components/TagComponents/TagGroupContainer'
 import { TagLabel } from '@/components/TagLabel'
 
-type CalendarProps = {
-  startDate: string
-}
+import { Calendar } from './CalendarCards/Calendar'
 
 type PlaceholderProps = {
   text: string
@@ -51,21 +49,6 @@ function Placeholder({ text }: PlaceholderProps) {
   return (
     <div className="py-8 lg:flex lg:h-[558px] lg:items-center lg:justify-center">
       {text}
-    </div>
-  )
-}
-
-function Calendar({ startDate }: CalendarProps) {
-  const { day, month } = formatDateComponentsFromISO(startDate)
-
-  return (
-    <div className="grid h-40 w-full grid-rows-[_40px,auto] rounded-md border border-blue-400 bg-brand-500 sm:w-[140px]">
-      <span className="flex items-center justify-center rounded-t text-base font-normal uppercase">
-        {month}
-      </span>
-      <span className="flex items-center justify-center rounded-b bg-brand-100 text-5xl font-bold text-brand-400">
-        {day}
-      </span>
     </div>
   )
 }
